feat: altera tipo de ID de Long para Integer nas cotas
- Modifica o tipo de ID nas classes CotaController, CotaService e CotaRepository de Long para Integer, garantindo consistência no tratamento de identificadores.
- Essas alterações visam simplificar a manipulação de IDs e melhorar a integração com o banco de dados, mantendo a integridade das operações de CRUD.

+++ src/main/java/controller/CotaController.java
@@ -52,7 +52,7 @@ public class CotaController {
     }
 
     @GetMapping("/{id}/editar")
-    public String editarCota(@PathVariable Long id, Model model) {
+    public String editarCota(@PathVariable Integer id, Model model) {
         try {
             model.addAttribute("cota", cotaService.buscarPorId(id));
             model.addAttribute("entidades", entidadeService.listarTodos());
@@ -75,7 +75,7 @@ public class CotaController {
 
     @DeleteMapping("/{id}")
     @ResponseBody
-    public String deletarCota(@PathVariable Long id) {
+    public String deletarCota(@PathVariable Integer id) {
         try {
             cotaService.deletar(id);
             return "Cota excluída com sucesso";
@@ -85,7 +85,7 @@ public class CotaController {
     }
 
     @GetMapping("/{id}/visualizar")
-    public String visualizar(@PathVariable Long id, Model model) {
+    public String visualizar(@PathVariable Integer id, Model model) {
         try {
             model.addAttribute("cota", cotaService.buscarPorId(id));
             return "cota/visualizar";
